# perf(git): cache git exec results