Cache results of the `git exec calls`.

Pretty anecdotal/not-profile-driven perf improvements, but intuitively it makes sense anyways.

For example, on a project i'm working on I get over 200 of:
```
Calling git with args 'rev-parse --show-toplevel'
```
with > 700 git commands executing

before:
```
❯ time gfs validate
garden validate 42.44s user 63.68s system 237% cpu 44.720 total
```

after:
```
❯ time gfs validate                   
garden validate  40.08s user 52.55s system 279% cpu 33.179 total
```

**What this PR does / why we need it**:

Faster commands :)